# ai-version/openapi.json
{
  "openapi": "3.0.0",
  "info": { "title": "Auth API (AI version)", "version": "1.0.0" },
  "servers": [{ "url": "http://localhost:3001" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "JWT" }
    }
  },
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" }, "400": { "description": "Bad Request" } }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Log in and receive a JWT",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK" }, "400": { "description": "Bad Request" }, "401": { "description": "Unauthorized" } }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "Log out",
        "security": [{ "bearerAuth": [] }],
        "responses": { "204": { "description": "No Content" }, "401": { "description": "Unauthorized" } }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Get the current user's profile",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" }, "401": { "description": "Unauthorized" } }
      }
    },
    "/protected/dashboard": {
      "get": {
        "summary": "Get the current user's dashboard",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "OK" }, "401": { "description": "Unauthorized" } }
      }
    },
    "/public/info": {
      "get": {
        "summary": "Public information",
        "responses": { "200": { "description": "OK" } }
      }
    }
  }
}